pyproject.toml
[project]
name = "transit-simulator"
version = "0.1.0"
description = ""
authors = [
    {name = "Maharan Eiffel Murshed",email = "maharan.murshed@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "alembic (>=1.14.1,<2.0.0)",
    "pandas (>=3.0.0,<4.0.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "shapely (>=2.1.2,<3.0.0)",
    "pyproj (>=3.7.2,<4.0.0)",
    "folium (>=0.20.0,<0.21.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "protobuf (>=6.33.5,<7.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]

[tool.poetry]
packages = [
    { include = "core", from = "src" },
    { include = "hamburg_twin", from = "src" },
    { include = "infra", from = "src" },
    { include = "rt", from = "src" }
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.15.1,<0.16.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]
